api/requirements.txt
fastapi
uvicorn
redis
yt-dlp
ffmpeg-python
boto3
pydantic-settings